Removed unnecessary files, imports, and moved mappers to separate package
