--- app/src/main/java/com/peeranm/simpledictionary/feature_word_meaning/utils/DummyData.kt
@@ -1,35 +1,0 @@
-package com.peeranm.simpledictionary.feature_word_meaning.utils
-
-val dummyData: Map<String, List<String>> = mutableMapOf<String, List<String>>().apply {
-    put(
-        "CRICKET_TEAMS",
-        listOf(
-            "India",
-            "Pakistan",
-            "Australia",
-            "England",
-            "South Africa"
-        )
-    )
-    put(
-        "FOOTBALL_TEAMS",
-        listOf(
-            "Brazil",
-            "Spain",
-            "Germany",
-            "Netherlands",
-            "Italy"
-        )
-    )
-    put(
-        "BASKETBALL_TEAMS",
-        listOf(
-            "United States",
-            "Spain",
-            "Argentina",
-            "France",
-            "Russia"
-        )
-    )
-}
-
